Fix bugs in controller
